remove circular dependency in practiceView
