+++ ui/analyse/src/practice/practiceView.ts
@@ -3,7 +3,7 @@ import { h, VNode } from 'snabbdom';
 import { bind, MaybeVNodes } from 'common/snabbdom';
 import { PracticeCtrl, Comment } from './practiceCtrl';
 import AnalyseCtrl from '../ctrl';
-import { renderNextChapter } from '../view';
+import { renderNextChapter } from '../study/nextChapter';
 
 function commentBest(c: Comment, root: AnalyseCtrl, ctrl: PracticeCtrl): MaybeVNodes {
   return c.best
